# Conflicts:
#	indra/llcommon/llqueuedthread.cpp
#	indra/llcommon/llsdserialize.cpp
#	indra/llcommon/llsdserialize.h
#	indra/llcommon/llthreadsafequeue.h
#	indra/llinventory/llsettingssky.cpp
#	indra/llui/llfolderviewitem.cpp
#	indra/llui/llview.h
#	indra/newview/llaisapi.cpp
#	indra/newview/llappviewer.cpp
#	indra/newview/llfloaterproperties.cpp
#	indra/newview/llfloaterproperties.h
#	indra/newview/llinventorybridge.cpp
#	indra/newview/llinventoryfunctions.cpp
#	indra/newview/llinventorypanel.cpp
#	indra/newview/llpanelmaininventory.cpp
#	indra/newview/llpanelmaininventory.h
#	indra/newview/llpanelmarketplaceinbox.cpp
#	indra/newview/llpanelprofile.cpp
#	indra/newview/llpanelprofile.h
#	indra/newview/llpathfindingnavmesh.cpp
#	indra/newview/llselectmgr.cpp
#	indra/newview/llsidepanelinventory.cpp
#	indra/newview/llsidepanelinventory.h
#	indra/newview/llsidepaneliteminfo.cpp
#	indra/newview/llsidepaneltaskinfo.cpp
#	indra/newview/llviewerfloaterreg.cpp
#	indra/newview/skins/default/xui/da/floater_inventory_item_properties.xml
#	indra/newview/skins/default/xui/de/floater_inventory_item_properties.xml
#	indra/newview/skins/default/xui/en/floater_inventory_item_properties.xml
#	indra/newview/skins/default/xui/en/menu_inventory_gear_default.xml
#	indra/newview/skins/default/xui/en/panel_main_inventory.xml
#	indra/newview/skins/default/xui/en/panel_profile_firstlife.xml
#	indra/newview/skins/default/xui/en/panel_profile_secondlife.xml
#	indra/newview/skins/default/xui/en/sidepanel_inventory.xml
#	indra/newview/skins/default/xui/en/sidepanel_item_info.xml
#	indra/newview/skins/default/xui/en/sidepanel_task_info.xml
#	indra/newview/skins/default/xui/es/floater_inventory_item_properties.xml
#	indra/newview/skins/default/xui/fr/floater_inventory_item_properties.xml
#	indra/newview/skins/default/xui/it/floater_inventory_item_properties.xml
#	indra/newview/skins/default/xui/ja/floater_inventory_item_properties.xml
#	indra/newview/skins/default/xui/pl/floater_inventory_item_properties.xml
#	indra/newview/skins/default/xui/ru/floater_inventory_item_properties.xml

+++ indra/newview/llpathfindingnavmesh.cpp
@@ -145,7 +145,7 @@ void LLPathfindingNavMesh::handleNavMeshResult(const LLSD &pContent, U32 pNavMes
 			// std::string newStr(reinterpret_cast<const char *>(&value[0]), binSize);
 			// std::istringstream streamdecomp( newStr );
 			// </FS:Beq pp Rye> 
-			unsigned int decompBinSize = 0;
+			size_t decompBinSize = 0;
 			bool valid = false;
 			// <FS:Beq pp Rye> Use new variants of unzip_llsd calls
 			// U8* pUncompressedNavMeshContainer = unzip_llsdNavMesh( valid, decompBinSize, streamdecomp, binSize ) ;
